+++ src/Exception/ConnectionException.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * Copyright (C) Jyxon, Inc. All rights reserved.
+ * See LICENSE for license details.
+ */
+namespace Ulrack\Dbal\Pdo\Exception;
+
+use Exception;
+use Throwable;
+
+class ConnectionException extends Exception
+{
+    /**
+     * Constructor.
+     *
+     * @param Throwable $exception
+     */
+    public function __construct(Throwable $exception)
+    {
+        parent::__construct(
+            'Could not connect to database.',
+            (int) $exception->getCode()
+        );
+    }
+}