+++ tests/Component/Result/PdoQueryResultTest.php
@@ -4,15 +4,15 @@
  * See LICENSE for license details.
  */
 
-namespace Ulrack\Dbal\Pdo\Tests\Result;
+namespace Ulrack\Dbal\Pdo\Tests\Component\Result;
 
 use PHPUnit\Framework\TestCase;
-use Ulrack\Dbal\Pdo\Result\PdoQueryResult;
-use Ulrack\Dbal\Pdo\Result\PdoIterableQueryResult;
+use Ulrack\Dbal\Pdo\Component\Result\PdoQueryResult;
+use Ulrack\Dbal\Pdo\Component\Result\PdoIterableQueryResult;
 use PDOStatement;
 
 /**
- * @coversDefaultClass \Ulrack\Dbal\Pdo\Result\PdoQueryResult
+ * @coversDefaultClass \Ulrack\Dbal\Pdo\Component\Result\PdoQueryResult
  */
 class PdoQueryResultTest extends TestCase
 {