+++ CHANGELOG.md
@@ -4,6 +4,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0 - 2019-11-06
+### Added
+- `Ulrack\Dbal\Pdo\Exception\ConnectionException`
+
+### Changed
+- Unified the namespace conventions used across packages.
+
+### Deprecated
+- Nothing
+
+### Removed
+- Nothing
+
+### Fixed
+- Nothing
+
+### Security
+- Nothing
+
 ## 1.0.0 - 2019-05-01
 ### Added
 - `Ulrack\Dbal\Pdo\Connection\PdoConnection`
@@ -27,4 +46,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 - Nothing
 
-[Unreleased]: https://github.com/ulrack/dbal-pdo/compare/1.0.0...HEAD
+[Unreleased]: https://github.com/ulrack/dbal-pdo/compare/2.0.0...HEAD
+[2.0.0]: https://github.com/ulrack/dbal-pdo/compare/1.0.0...2.0.0